removed TrackerController
