+++ src/main/java/com/unb/tracker/web/CourseController.java
@@ -1,6 +1,7 @@
 package com.unb.tracker.web;
 
 import com.unb.tracker.model.Course;
+import com.unb.tracker.model.Seat;
 import com.unb.tracker.repository.CourseRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,9 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 
 import java.security.Principal;
 import java.util.List;
@@ -47,5 +46,31 @@ public class CourseController {
             return "instructor/course";
         }
         return "error";
+    }
+
+    @GetMapping(value="/courses/{courseId}")
+    public @ResponseBody
+    Course getCourse(@PathVariable Long courseId) {
+        return courseRepository.findOne(courseId);
+    }
+
+    @PostMapping(value="/courses/{courseId}/seats")
+    public @ResponseBody Course postCourseSeats(@PathVariable Long courseId, @RequestBody List<Seat> seats) {
+        Course course = courseRepository.findOne(courseId);
+        course.setSeats(seats);
+        courseRepository.save(course);
+        return course;
+    }
+
+    @PostMapping(value="/courses")
+    @ResponseBody
+    public String postCourse(@RequestBody Course course) {
+        courseRepository.save(course);
+        return "Saved";
+    }
+
+    @GetMapping(value="/courses")
+    public @ResponseBody Iterable<Course> getCourses() {
+        return courseRepository.findAll();
     }
 }
